materialToMD/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
firebase_admin==7.1.0
google-cloud-storage==3.8.0
google-cloud-firestore==2.23.0
sqlalchemy==2.0.45
pymupdf==1.26.7
python-docx==1.2.0
python-pptx==1.0.2
pillow==12.1.0
markdownify==1.2.2
beautifulsoup4==4.14.3
openpyxl==3.1.5
python-dotenv==1.2.1
pytest==9.0.2
httpx==0.28.1
requests==2.32.5

annotated-types==0.7.0
anyio==4.12.1
cachecontrol==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
et_xmlfile==2.0.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
rsa==4.9.1
six==1.17.0
soupsieve==2.8.1
starlette==0.50.0
typing_extensions==4.15.0
urllib3==2.6.3
xlsxwriter==3.2.9